backend/pyproject.toml
[project]
name = "broai-research-rocket"
version = "0.1.0"
description = "BroAI, your new research bro"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "broai==0.1.14",
    "celery[redis]>=5.5.2",
    "flower",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "redis>=5.3.0",
    "streamlit>=1.45.0",
    "torch==2.6.0",
    "uvicorn>=0.34.2",
]
